requirements.txt
# Core Dependencies for Crypto Trading Bot
pandas>=1.5.0
numpy>=1.24.0
asyncio

# Trading and Market Data
ccxt>=4.0.0
websocket-client>=1.6.0
python-binance>=1.0.19
yfinance>=0.2.0

# Technical Analysis
ta>=0.10.2
empyrical>=0.5.5

# Machine Learning
scikit-learn>=1.3.0
tensorflow>=2.13.0
xgboost>=1.7.0

# Data Storage and Caching
redis>=4.6.0
sqlalchemy>=2.0.0

# Web Framework and Dashboard
dash>=2.14.0
plotly>=5.15.0
flask>=2.3.0
streamlit>=1.25.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
schedule>=1.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
